+++ examples/middlewares-custom-decorators/index.ts
@@ -6,6 +6,7 @@ import { buildSchema } from "../../src";
 import { RecipeResolver } from "./recipe/recipe.resolver";
 import { ResolveTimeMiddleware } from "./middlewares/resolve-time";
 import { ErrorLoggerMiddleware } from "./middlewares/error-logger";
+import { Context } from "./context";
 
 async function bootstrap() {
   // build TypeGraphQL executable schema
@@ -16,7 +17,18 @@ async function bootstrap() {
   });
 
   // Create GraphQL server
-  const server = new ApolloServer({ schema });
+  const server = new ApolloServer({
+    schema,
+    context: (): Context => {
+      return {
+        // example user
+        currentUser: {
+          id: 123,
+          name: "Sample user",
+        },
+      };
+    },
+  });
 
   // Start the server
   const { url } = await server.listen(4000);