+++ examples/middlewares-custom-decorators/user.ts
@@ -1,0 +1,4 @@
+export default interface User {
+  id: number;
+  name: string;
+}