+++ examples/middlewares/context.ts
@@ -1,3 +1,0 @@
-export interface Context {
-  username?: string;
-}