+++ src/decorators/createMethodDecorator.ts
@@ -1,0 +1,8 @@
+import { UseMiddleware } from "./UseMiddleware";
+import { MiddlewareFn } from "../interfaces/Middleware";
+
+export function createMethodDecorator<TContextType = {}>(
+  resolver: MiddlewareFn<TContextType>,
+): MethodDecorator {
+  return UseMiddleware(resolver);
+}