+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelog and release notes
 
-<!-- ## Unreleased -->
+## Unreleased
 <!-- here goes all the unreleased changes descriptions -->
+## Features
+- add support for creating custom parameter decorators (#329)
+
 ## v0.17.3
 ### Features
 - update packages `semver` to `^6.0.0` and `graphql-subscriptions` to `^1.1.0`